refactor: split up components used in student timeline insights
